+++ src/test/java/calculator/domain/StringCalculatorTest.java
@@ -1,6 +1,5 @@
-package calculator;
+package calculator.domain;
 
-import calculator.domain.StringCalculator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
